refactor: handle stdout iteration with filter_map(Result::ok)
Clippy suggested replacing the previous code with `map_while(Result::ok)`,
which may be more concise but introduces a significant semantic change:
iteration stops at the first read error, truncating all subsequent output.

Since the goal is to display as much information as possible to the user
during the update process, a single I/O error should not block the rest of the log
from being shown. For this reason we use `filter_map(Result::ok)`.